move current decoder to bit_stream_decoder
